requirements.txt
cleverbot
discord.py